+++ abel/linbasex.py
@@ -7,6 +7,7 @@ from __future__ import unicode_literals
 import numpy as np
 import abel
 import os
+from glob import glob
 import scipy
 from scipy.special import eval_legendre
 from scipy import ndimage
@@ -471,8 +472,10 @@ def get_bs_cached(cols, basis_dir=None, legendre_orders=[0, 2],
     cols : int
         width of image
 
-    basis_dir : str
-        path to the directory for saving / loading the basis
+    basis_dir : str or None
+        path to the directory for saving / loading the basis. Use ``''`` for
+        the default directory. If ``None``, the basis set will not be loaded
+        from or saved to disk.
 
     legendre_orders : list
         default [0, 2] = 0 order and 2nd order polynomials
@@ -526,6 +529,8 @@ def get_bs_cached(cols, basis_dir=None, legendre_orders=[0, 2],
     _pas = pas
     _radial_step = radial_step
     _clip = clip
+    if basis_dir == '':
+        basis_dir = abel.transform.get_basis_dir(make=True)
     if basis_dir is not None:
         path_to_basis_file = os.path.join(basis_dir, basis_name)
         if os.path.exists(path_to_basis_file):
@@ -576,6 +581,33 @@ def cache_cleanup():
     _pas = None
     _radial_step = None
     _clip = None
+
+
+def basis_dir_cleanup(basis_dir=''):
+    """
+    Utility function.
+
+    Deletes basis sets saved on disk.
+
+    Parameters
+    ----------
+    basis_dir : str or None
+        relative or absolute path to the directory with saved basis sets. Use
+        ``''`` for the default directory. ``None`` does nothing.
+
+    Returns
+    -------
+    None
+    """
+    if basis_dir == '':
+        basis_dir = abel.transform.get_basis_dir(make=False)
+
+    if basis_dir is None:
+        return
+
+    files = glob(os.path.join(basis_dir, 'linbasex_basis_*.npy'))
+    for fname in files:
+        os.remove(fname)
 
 
 linbasex_transform.__doc__ += _linbasex_parameter_docstring