+++ examples/example_basex_step.py
@@ -27,10 +27,8 @@ right_half = st.abel[center:]
 left_half = st.abel[:center+1][::-1]
 # BASEX Transform:
 # Calculate the inverse abel transform for the centered data
-recon_right = basex_transform(
-    right_half, basis_dir='bases', dr=st.dr, verbose=True)
-recon_left = basex_transform(
-    left_half, basis_dir='bases', dr=st.dr, verbose=False)
+recon_right = basex_transform(right_half, dr=st.dr, verbose=True)
+recon_left = basex_transform(left_half, dr=st.dr, verbose=False)
 plt.plot(
     st.r[center:], recon_right, '--.', c='red',
     label='Inverse transform [BASEX]')