+++ examples/example_simple_GUI.py
@@ -114,8 +114,7 @@ def _transform():
                              transform_options=dict(return_Beta=True))
     else:
         AIM = abel.Transform(IM, method=method, direction="inverse",
-                             symmetry_axis=None,
-                             transform_options=dict(basis_dir='bases'))
+                             symmetry_axis=None)
 
     f.clf()
     a = f.add_subplot(111)